Add solved and constructed sudokus
I wanted to publish the sudokus I created somewhere in addition to
tracking my progress for solving sudokus.  Why not publish that on here,
that's what this would be for right?  I probably won't push an update
every time I solve a sudoku, just when I have a few saved up.

Signed-off-by: anderium <33520919+anderium@users.noreply.github.com>