Add lambda calculus section
It's been months since I did this, so I'm quite sure I won't be adding
to this any time soon, which is why I'll just publish it. I think the
final section is unfinished, but I'm not sure. Either way this is just a
link to a nice interpreter, actually.

By the way, I have a bug fix open for it, which tells me that it's been
three quarter years since I did all this…

Signed-off-by: anderium <33520919+anderium@users.noreply.github.com>